+++ src/node/_jwks_cert_b.ts
@@ -1,8 +1,10 @@
 import { _b } from '@ctx-core/object'
 import { _jwks_x5c_b, _jwks_x5c_ctx_I } from './_jwks_x5c_b'
-export const _jwks_cert_b = _b('_jwks_cert', (
-	ctx:_jwks_cert_ctx_I
-)=>{
+const key = '_jwks_cert'
+export interface _jwks_cert_ctx_I extends _jwks_x5c_ctx_I {
+	_jwks_cert?:_jwks_cert_T
+}
+export const _jwks_cert_b = _b<_jwks_cert_ctx_I, typeof key>(key, ctx=>{
 	const _jwks_x5c = _jwks_x5c_b(ctx)
 	return _jwks_cert as _jwks_cert_T
 	async function _jwks_cert() {
@@ -16,7 +18,4 @@ export const _jwks_cert_b = _b('_jwks_cert', (
 		return jwks_cert
 	}
 })
-export interface _jwks_cert_ctx_I extends _jwks_x5c_ctx_I {
-	_jwks_cert?:_jwks_cert_T
-}
 export type _jwks_cert_T = ()=>Promise<string>